refactor: :recycle: Refactor TextInput. Add forwardRef
