+++ libs/dana/src/components/TextInput/TextInput.stories.tsx
@@ -1,8 +1,7 @@
 import { TextInput, TextInputProps } from './TextInput';
-import { asPlayground, asChromaticStory } from '../../../lib/story-intents';
-import type { Story } from '../../../lib/storybook-emotion-10-fixes';
+import { asPlayground, asChromaticStory } from '../../lib/story-intents';
+import type { Story } from '../../lib/storybook-emotion-10-fixes';
 import { useState } from 'react';
-import { ArrowRightIcon } from '@dana-icons';
 
 export default {
     title: 'Components/Forms/Text Input',
