+++ libs/dana/src/components/forms/index.ts
@@ -1,5 +1,4 @@
 export * from './number-input';
-export * from './text-input';
 export * from './button';
 export * from './form';
 export * from './input';
